Significant Update to DFW API and DFWS Features
This is a significant update adding dialogue and a number of
features/events to the Devious Framework Support (DFWS) mod's leash
game.
To facilitate these changes a number of changes and additions were also
needed in the Devious Framework (DFW) API, including the feature of
remembering NPC dispositions.
Additionally locking the player in BDSM furniture has been added to the
DFWS mod.  Corresponding support has been added to the Devious Framework
as well.