Detach Initialization From MCM
Separated the main script initialization code from the MCM
initialization process.  The main script initialization for both mods is
now run in the context of the first update event.  I have a suspicion
that running it in the MCM initialization context was sometimes causing
deadlock process dependencies.
Fixed a couple of other bugs of little consequence.