Cleanups Post 2.01 Release
Added: Simple Slavery as an outcome of the leash game (very basic for
now).
Made (some) dialogue response actions faster and more responsive.
Cleaned up a couple unfinished dialogues that did not end properly.
Changed: Leash Game to Furniture transfer now happens at the end of the
leash game only.
Added: the ability to assign specific cells to regions.
Fixed: some farm interiors not allocated to the Windhelm region.
Fixed: Do not call for help if menus (or the console) are open.
Fixed: a race condition cleaning the nearby actor list.
Added: The ability to clear a Master based on the mod string in case the
Master is None.
API Changed: GetRegion() and ClearMaster()